Lecture 4: 1D Inviscid Burgers' Equation
Numerically and "analytically" (method of characteristics) solve 1D
Inviscid Burgers' Equation.
Show that we get a steepening of the velocity (u(x)) in the x-domain due
to unequal accelerations in the x-domain.
Show that viscosity prevents formation of shock waves.
Show that dispersive numerical schemes propagate shock at different
speeds.