Add Lecture 13 notebook.
Forced the vertex IDs to create this transect. Didn't really use
sections.py package to make calculations. Need to figure out why certain
transects never reached their end point... May need to write code that
breaks the transect algorithm if the xr_end_vertex is one of the
potential vertices you could move to, even if it is not the closest.